+++ __tests__/services/grpc/state_serialization_client_grpc.service.spec.ts
@@ -43,7 +43,39 @@ describe('credentials', () => {
 });
 
 describe('getSessionState', () => {
-  test('should return session state', async () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
+  test('should return full session state', async () => {
+    const interactionId = v4();
+    const sessionState = new SessionState()
+      .setState(v4())
+      .setCreationTime(protoTimestamp())
+      .setVersion(new SessionState.Version().setInteractionId(interactionId));
+    const generateSessionToken = jest
+      .spyOn(StateSerializationClient.prototype, 'getSessionState')
+      .mockImplementationOnce((_request, _metadata, _options, callback) => {
+        const cb = typeof _options === 'function' ? _options : callback;
+        cb(null, sessionState);
+        return {} as SurfaceCall;
+      });
+
+    const service = new StateSerializationClientGrpcService();
+    const result = await service.getSessionState({
+      scene: SCENE,
+      sessionToken,
+    });
+
+    expect(generateSessionToken).toHaveBeenCalledTimes(1);
+    expect(result.state).toEqual(sessionState.getState_asB64());
+    expect(result.creationTime).toEqual(
+      sessionState.getCreationTime().toDate().toISOString(),
+    );
+    expect(result.version?.interactionId).toEqual(interactionId);
+  });
+
+  test('should return session state without version', async () => {
     const sessionState = new SessionState()
       .setState(v4())
       .setCreationTime(protoTimestamp());
@@ -66,5 +98,6 @@ describe('getSessionState', () => {
     expect(result.creationTime).toEqual(
       sessionState.getCreationTime().toDate().toISOString(),
     );
+    expect(result.version).toBeUndefined();
   });
 });