+++ src/index.ts
@@ -72,6 +72,7 @@ import {
   FeedbackDislikeProps,
   FeedbackLikeProps,
 } from './services/feedback.service';
+import { SessionState } from './services/gprc/state_serialization_client_grpc.service';
 import { InworldConnectionService } from './services/inworld_connection.service';
 
 const config = Config.getInstance();
@@ -120,6 +121,7 @@ export {
   Session,
   SessionContinuation,
   SessionContinuationProps,
+  SessionState,
   SessionToken,
   SessionTokenProps,
   status,