+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add getProto method to Inworld packets
 - Support SPEECH_RECOGNITION_ONLY mode
+- Add version to session state
 
 ## [1.14.2] - 2024-06-27